move slow test framework integration tests to separate jobs (#2955)
